intro lattice  project for tree pricers, and zip iterating and lattice sampling utils
