+++ Vectorisation.sln
@@ -26,6 +26,11 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "cumNormalExample", "cumNorm
 		{837BEACF-BCED-4AB2-BAF5-A7254EB6756B} = {837BEACF-BCED-4AB2-BAF5-A7254EB6756B}
 	EndProjectSection
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "lattice", "lattice\lattice.vcxproj", "{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}"
+	ProjectSection(ProjectDependencies) = postProject
+		{837BEACF-BCED-4AB2-BAF5-A7254EB6756B} = {837BEACF-BCED-4AB2-BAF5-A7254EB6756B}
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		clang-cl|ARM64 = clang-cl|ARM64
@@ -220,6 +225,34 @@ Global
 		{503BFAC4-7F7F-4B0F-A09D-A9985A9382D0}.Release|x64.Build.0 = Release|x64
 		{503BFAC4-7F7F-4B0F-A09D-A9985A9382D0}.Release|x86.ActiveCfg = Release|Win32
 		{503BFAC4-7F7F-4B0F-A09D-A9985A9382D0}.Release|x86.Build.0 = Release|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clang-cl|ARM64.ActiveCfg = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clang-cl|ARM64.Build.0 = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clang-cl|x64.ActiveCfg = Debug|x64
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clang-cl|x64.Build.0 = Debug|x64
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clang-cl|x86.ActiveCfg = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clang-cl|x86.Build.0 = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clangD|ARM64.ActiveCfg = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clangD|ARM64.Build.0 = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clangD|x64.ActiveCfg = Debug|x64
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clangD|x64.Build.0 = Debug|x64
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clangD|x86.ActiveCfg = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.clangD|x86.Build.0 = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.Debug|ARM64.ActiveCfg = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.Debug|x64.ActiveCfg = Debug|x64
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.Debug|x64.Build.0 = Debug|x64
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.Debug|x86.ActiveCfg = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.Debug|x86.Build.0 = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.ICC2022|ARM64.ActiveCfg = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.ICC2022|ARM64.Build.0 = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.ICC2022|x64.ActiveCfg = Debug|x64
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.ICC2022|x64.Build.0 = Debug|x64
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.ICC2022|x86.ActiveCfg = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.ICC2022|x86.Build.0 = Debug|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.Release|ARM64.ActiveCfg = Release|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.Release|x64.ActiveCfg = Release|x64
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.Release|x64.Build.0 = Release|x64
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.Release|x86.ActiveCfg = Release|Win32
+		{FFE7CCAE-431C-44CD-8C12-80BF4B81ED15}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
